requirements.txt
# Polymarket CLOB Client
py-clob-client>=0.34.5

# Ethereum and Web3
eth-account>=0.8.0
eth-utils>=2.1.0
web3>=6.0.0

# HTTP and Async
aiohttp>=3.8.0
requests>=2.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Environment Variables
python-dotenv>=0.21.0

# Logging
coloredlogs>=15.0

# Testing
pytest>=7.2.0
pytest-asyncio>=0.20.0
